ESP32_Distance_Tracker/pc_monitor/requirements.txt
pyserial>=3.5
matplotlib>=3.5.0
pandas>=1.3.0
colorama>=0.4.4
